requirements.txt
flask==3.0.0
flask-cors==4.0.0
together==1.2.0
chromadb==0.4.22
python-dotenv==1.0.0
pymupdf==1.23.8
gunicorn==21.2.0
